requirements.txt
authlib==1.6.3
fastapi==0.117.1
geoalchemy2==0.18.0
ijson==3.4.0
numpy==2.3.3
passlib==1.7.4
pydantic==2.11.9
pydantic_settings==2.11.0
python_jose==3.3.0
Shapely==2.1.2
SQLAlchemy==2.0.29
sqlmodel==0.0.25
starlette==0.48.0
python-dotenv==1.0.1
uvicorn==0.29.0
asyncpg==0.29.0
itsdangerous==2.1.2
python-multipart==0.0.9
passlib[bcrypt]
pyproj
py-wake 
topfarm 
pyproj 
xarray 
pandas 
matplotlib